Setting type of study_id and study_file_id to ObjectId as needed
In order to preserve associations in Mongo, we need to set all "_id" attributes to be a BSON ObjectId rather than a string.  Results returned from inserts/queries will already be in the correct format, but args passed to ingest on the command line are strings and therefore need to be cast.